+++ src/debug/websocket/subscriptions/price.js
@@ -1,0 +1,21 @@
+
+
+// const SUBSCRIPTION_QUERY = gql`
+// subscription Subscription($input: LnInvoicePaymentStatusInput!) {
+//   	lnInvoicePaymentStatus(input: $input) {
+//     		errors {
+//       		message
+//     	}
+//     		status
+//   	}
+// }
+// `;
+
+// export const priceUpdatesGql = (pr) => ({
+//     query: SUBSCRIPTION_QUERY,
+//     variables: {
+//         input: {
+//             "paymentRequest": pr
+//         }
+//     }
+// })
